+++ src/plugins/ramps/allRampPlugins.ts
@@ -1,5 +1,7 @@
 import { banxaRampPlugin } from './banxa/banxaRampPlugin'
+import { bitsofgoldRampPlugin } from './bitsofgold/bitsofgoldRampPlugin'
 import { infiniteRampPlugin } from './infinite/infiniteRampPlugin'
+import { libertyxRampPlugin } from './libertyx/libertyxRampPlugin'
 import { moonpayRampPlugin } from './moonpay/moonpayRampPlugin'
 import { paybisRampPlugin } from './paybis/paybisRampPlugin'
 import type { RampPluginFactory } from './rampPluginTypes'
@@ -8,7 +10,9 @@ import { simplexRampPlugin } from './simplex/simplexRampPlugin'
 
 export const pluginFactories: Record<string, RampPluginFactory> = {
   banxa: banxaRampPlugin,
+  bitsofgold: bitsofgoldRampPlugin,
   infinite: infiniteRampPlugin,
+  libertyx: libertyxRampPlugin,
   moonpay: moonpayRampPlugin,
   paybis: paybisRampPlugin,
   revolut: revolutRampPlugin,