+++ src/plugins/ramps/bitsofgold/bitsofgoldRampTypes.ts
@@ -1,0 +1,12 @@
+import { asObject, asOptional, asString } from 'cleaners'
+
+export interface InitOptions {
+  readonly partnerIcon: string
+}
+
+export const asInitOptions = asObject<InitOptions>({
+  partnerIcon: asOptional(
+    asString,
+    `https://content.edge.app/bits-of-gold-logo.png`
+  )
+}).withRest